Add test for react component about dom
