+++ app/assets/javascripts/components/_navbar.js
