+++ app/assets/javascripts/components/admin/_admin_menu.js
