+++ app/assets/javascripts/components/admin/actors/detail/admin_actor.js
@@ -1,10 +1,10 @@
 import React, { Component, PropTypes } from 'react'
-import AdminActorDetail from './admin_actor_detail.js'
+import AdminActorDetail from './_admin_actor_detail.js'
 
 export default class AdminActor extends Component {
   render() {
     return (
-      <div>
+      <div className='adminActorComponent'>
         <AdminActorDetail url={'api/admin/actors/' + this.props.params.actorId} />
       </div>
     )
