+++ app/assets/javascripts/components/admin/actors/list/_admin_actor_row.js
