+++ app/assets/javascripts/components/admin/actors/list/admin_actors.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react'
-import AdminActorsTable from './admin_actors_table.js'
+import AdminActorsTable from './_admin_actors_table.js'
 
 export default class AdminActors extends Component {
   render() {
     return (
-      <div>
+      <div className='adminActorsComponent'>
         <AdminActorsTable url='api/admin/actors' />
       </div>
     )
