+++ app/assets/javascripts/components/admin/admin.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react'
-import AdminMenu from './admin_menu.js'
+import AdminMenu from './_admin_menu.js'
 
 export default class Admin extends Component {
   render() {
