+++ app/assets/javascripts/components/admin/animes/detail/admin_anime.js
@@ -1,10 +1,10 @@
 import React, { Component, PropTypes } from 'react'
-import AdminAnimeDetail from './admin_anime_detail.js'
+import AdminAnimeDetail from './_admin_anime_detail.js'
 
 export default class AdminAnime extends Component {
   render() {
     return (
-      <div>
+      <div className='adminAnimeComponent'>
         <AdminAnimeDetail url={'api/admin/animes/' + this.props.params.animeId} />
       </div>
     )
