+++ app/assets/javascripts/components/admin/animes/list/_admin_anime_row.js
