+++ app/assets/javascripts/components/app.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from 'react'
-import Navbar from './navbar.js'
+import Navbar from './_navbar.js'
 
 export default class App extends Component {
   render() {
