+++ app/assets/javascripts/components/welcome/_anime.js
