+++ app/assets/javascripts/routes.js
@@ -6,11 +6,11 @@ import Welcome from './components/welcome/welcome.js'
 import Admin from './components/admin/admin.js'
 import AdminTopPage from './components/admin/top/admin_top_page.js'
 import AdminAnimesPage from './components/admin/animes/admin_animes_page.js'
-import AdminAnimes from './components/admin/animes/admin_animes.js'
-import AdminAnime from './components/admin/animes/admin_anime.js'
+import AdminAnimes from './components/admin/animes/list/admin_animes.js'
+import AdminAnime from './components/admin/animes/detail/admin_anime.js'
 import AdminActorsPage from './components/admin/actors/admin_actors_page.js'
-import AdminActors from './components/admin/actors/admin_actors.js'
-import AdminActor from './components/admin/actors/admin_actor.js'
+import AdminActors from './components/admin/actors/list/admin_actors.js'
+import AdminActor from './components/admin/actors/detail/admin_actor.js'
 
 module.exports = (
   <Route component={App} path="/">
